Added a text and quad manager
