+++ StortSpelProjekt/Engine/src/Engine.h
@@ -26,6 +26,8 @@
 #include "Renderer/HeightmapModel.h"
 #include "Renderer/Mesh.h"
 #include "Renderer/BaseCamera.h"
+#include "Renderer/TextManager.h"
+#include "Renderer/QuadManager.h"
 #include "Physics/Physics.h"
 #include "AudioEngine/AudioBuffer.h"
